Device snapshotter
