+++ paparazzi/src/main/java/app/cash/paparazzi/snapshotter/Clip.kt
@@ -1,0 +1,9 @@
+package app.cash.paparazzi.snapshotter
+
+import java.awt.image.BufferedImage
+import kotlinx.coroutines.flow.Flow
+
+data class Clip(
+  val spec: ClipSpec,
+  val images: Flow<BufferedImage>
+)
