+++ paparazzi/src/main/java/app/cash/paparazzi/snapshotter/ClipSpec.kt
@@ -1,0 +1,8 @@
+package app.cash.paparazzi.snapshotter
+
+data class ClipSpec(
+  val frameSpec: FrameSpec,
+  val start: Long,
+  val end: Long,
+  val fps: Int
+)
