+++ paparazzi/src/main/java/app/cash/paparazzi/snapshotter/Snapshot.kt
@@ -1,0 +1,8 @@
+package app.cash.paparazzi.snapshotter
+
+import java.awt.image.BufferedImage
+
+data class Snapshot(
+  val spec: FrameSpec,
+  val image: BufferedImage
+)
