+++ paparazzi/src/main/java/app/cash/paparazzi/test/TestName.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.paparazzi
+package app.cash.paparazzi.test
 
 data class TestName(
   val packageName: String,
