+++ paparazzi/src/test/java/app/cash/paparazzi/HtmlReportWriterTest.kt
@@ -16,6 +16,12 @@
 package app.cash.paparazzi
 
 import app.cash.paparazzi.FileSubject.Companion.assertThat
+import app.cash.paparazzi.snapshotter.FrameSpec
+import app.cash.paparazzi.snapshotter.Snapshot
+import app.cash.paparazzi.test.HtmlReportWriter
+import app.cash.paparazzi.test.TestName
+import app.cash.paparazzi.test.TestRecord
+import app.cash.paparazzi.test.sanitizeForFilename
 import com.google.common.truth.Truth.assertThat
 import org.junit.Rule
 import org.junit.Test
@@ -42,19 +48,18 @@ class HtmlReportWriterTest {
   fun happyPath() {
     val htmlReportWriter = HtmlReportWriter("run_one", reportRoot.root)
     htmlReportWriter.use {
-      val frameHandler = htmlReportWriter.newFrameHandler(
+      htmlReportWriter.handleSnapshot(
         Snapshot(
+          FrameSpec(),
+          anyImage
+        ),
+        TestRecord(
           name = "loading",
           testName = TestName("app.cash.paparazzi", "CelebrityTest", "testSettings"),
           timestamp = Instant.parse("2019-03-20T10:27:43Z").toDate(),
           tags = listOf("redesign")
-        ),
-        1,
-        -1
+        )
       )
-      frameHandler.use {
-        frameHandler.handle(anyImage)
-      }
     }
 
     assertThat(File("${reportRoot.root}/index.js")).hasContent(
@@ -90,28 +95,6 @@ class HtmlReportWriterTest {
   }
 
   @Test
-  fun noSnapshotOnFailure() {
-    val htmlReportWriter = HtmlReportWriter("run_one", reportRoot.root)
-    htmlReportWriter.use {
-      val frameHandler = htmlReportWriter.newFrameHandler(
-        snapshot = Snapshot(
-          name = "loading",
-          testName = TestName("app.cash.paparazzi", "CelebrityTest", "testSettings"),
-          timestamp = Instant.parse("2019-03-20T10:27:43Z").toDate()
-        ),
-        frameCount = 4,
-        fps = -1
-      )
-      frameHandler.use {
-        // intentionally empty, to simulate no content written on exception
-      }
-    }
-
-    assertThat(File(reportRoot.root, "images")).isEmptyDirectory()
-    assertThat(File(reportRoot.root, "videos")).isEmptyDirectory()
-  }
-
-  @Test
   fun alwaysOverwriteOnRecord() {
     // set record mode
     System.setProperty("paparazzi.test.record", "true")
@@ -119,7 +102,7 @@ class HtmlReportWriterTest {
     val htmlReportWriter = HtmlReportWriter("record_run", reportRoot.root, snapshotRoot.root)
     htmlReportWriter.use {
       val now = Instant.parse("2021-02-23T10:27:43Z")
-      val snapshot = Snapshot(
+      val testRecord = TestRecord(
         name = "test",
         testName = TestName("app.cash.paparazzi", "HomeView", "testSettings"),
         timestamp = now.toDate()
@@ -130,13 +113,13 @@ class HtmlReportWriterTest {
       // precondition
       assertThat(golden).doesNotExist()
 
-      // take 1
-      val frameHandler1 = htmlReportWriter.newFrameHandler(
-        snapshot = snapshot,
-        frameCount = 1,
-        fps = -1
+      htmlReportWriter.handleSnapshot(
+        Snapshot(
+          FrameSpec(),
+          anyImage
+        ),
+        testRecord
       )
-      frameHandler1.use { frameHandler1.handle(anyImage) }
       assertThat(golden).exists()
       val timeFirstWrite = golden.lastModifiedTime()
 
@@ -144,12 +127,13 @@ class HtmlReportWriterTest {
       Thread.sleep(100)
 
       // take 2
-      val frameHandler2 = htmlReportWriter.newFrameHandler(
-        snapshot = snapshot.copy(timestamp = now.plusSeconds(1).toDate()),
-        frameCount = 1,
-        fps = -1
+      htmlReportWriter.handleSnapshot(
+        Snapshot(
+          FrameSpec(),
+          anyImage
+        ),
+        testRecord.copy(timestamp = now.plusSeconds(1).toDate())
       )
-      frameHandler2.use { frameHandler2.handle(anyImage) }
       assertThat(golden).exists()
       val timeOverwrite = golden.lastModifiedTime()
